🦄 refactor(rename): email
